revalidate command, added support for a dry run
